Converted random effects to spatially explicit random effects.
Most of the problems of estimating site-level variance have gone away, so this looks like it's better. Also reveals a bunch of spatial patterns, likely tied to habitat prefs. Neat!

Could a version of this be used for the different habitat types? i.e. weighting distances from different habitat features to use in a regression?